Add checkbox on whether to include dataset as a node
