+++ package.json
@@ -18,9 +18,7 @@
     "svelte-preprocess-markdown": "^2.7.3"
   },
   "dependencies": {
-    "bootstrap": "^4.5.3",
     "sirv-cli": "^1.0.10",
-    "sveltestrap": "^3.11.0",
     "vis-data": "^7.1.1",
     "vis-network": "^8.5.5"
   }
